Добавила изменение эффектов изображения
